services/workers/tiktok-connector/pyproject.toml
[tool.poetry]
name = "tiktok-connector"
version = "0.1.0"
description = "AI Social Media Content Generator - TikTok Connector"
authors = ["Cursor AI <noreply@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
nats-py = "^2.7.2"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


